frontend: rename wwt.js to widget.js
