+++ frontend/lib/widget.js
